+++ webapp/.env.local.example
@@ -1,0 +1,3 @@
+# URL of the AWS Lambda endpoint that handles SSN + ZIP lookups.
+# Example: https://abc123.execute-api.us-east-1.amazonaws.com/prod/status
+LAMBDA_API_URL=
