+++ webapp/components/HorizontalDivider.tsx
@@ -1,0 +1,2 @@
+/** A simple horizontal rule divider. */
+export const HorizontalDivider = () => <div className="horizontal-divider" />;
