+++ webapp/global.d.ts
@@ -1,0 +1,4 @@
+declare module "*.css";
+declare module "*.png";
+declare module "*.svg";
+declare module "*.jpg";
