+++ webapp/package.json
@@ -13,8 +13,9 @@
   "license": "MIT",
   "author": "New Jersey State Office of Innovation",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
+    "dev": "next dev --port 3000",
+    "build": "next build",
+    "start": "next start",
     "test": "vitest run",
     "format": "prettier . --write",
     "format:check": "prettier . --check",
@@ -23,18 +24,16 @@
   "dependencies": {
     "@newjersey/njwds": "2.9.1",
     "@trussworks/react-uswds": "11.0.0",
-    "@vitejs/plugin-react": "6.0.2",
-    "jsdom": "29.1.1",
+    "next": "16.2.7",
     "react": "19.2.6",
     "react-dom": "19.2.6",
     "react-hook-form": "7.76.1",
-    "react-router-dom": "7.15.1",
     "typescript": "6.0.3"
   },
   "devDependencies": {
     "@testing-library/react": "16.3.2",
     "@types/node": "25.9.1",
-    "vite": "8.0.5",
+    "jsdom": "29.1.1",
     "vitest": "4.1.2"
   }
 }
