--- webapp/src/App.test.tsx
@@ -1,7 +1,0 @@
-import { describe, it, expect } from "vitest";
-
-describe("test", () => {
-  it("passes", () => {
-    expect(true).toBe(true);
-  });
-});
