--- webapp/src/components/HorizontalDivider.tsx
@@ -1,3 +1,0 @@
-export const HorizontalDivider = () => {
-  return <div className="horizontal-divider" />;
-};
