--- webapp/src/main.tsx
@@ -1,16 +1,0 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-
-import App from "./App";
-
-const rootElement = document.getElementById("root");
-
-if (!rootElement) {
-  throw new Error("Root mount point not found. Expected a <div id='root'> element in index.html.");
-}
-
-createRoot(rootElement).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-);
