--- webapp/vite.config.ts
@@ -1,12 +1,0 @@
-/// <reference types="vitest/config" />
-import react from "@vitejs/plugin-react";
-import { defineConfig } from "vite";
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  resolve: {},
-  server: {
-    port: 3000,
-  },
-});
