--- webapp/vitest.setup.ts
@@ -1,1 +1,0 @@
-// Vitest global setup — add test utilities and matchers here as needed.
